pyproject.toml
[project]
name = "airflow-projeto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.0.2",
    "boto3>=1.39.4",
    "pandas>=2.3.1",
    "requests>=2.32.4",
]
